+++ .storybook/stories/TableFiltering/components/DashboardFilterModal.tsx
@@ -1,123 +1,16 @@
-import { Box, Stack } from '@ag.ds-next/react/box';
+import { Stack } from '@ag.ds-next/react/box';
 import { GetDataFilters } from '../lib/getData';
 import { DateRangePicker } from '@ag.ds-next/react/date-picker';
-import { Button, ButtonGroup } from '@ag.ds-next/react/button';
 import { FilterStatusSelect } from './FilterStatusSelect';
-import {
-	Fragment,
-	PropsWithChildren,
-	ReactNode,
-	useEffect,
-	useRef,
-	useState,
-} from 'react';
+import { Fragment, useEffect, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { Global } from '@emotion/react';
-import FocusLock from 'react-focus-lock';
-import { animated, useSpring } from '@react-spring/web';
-import {
-	mapSpacing,
-	tokens,
-	usePrefersReducedMotion,
-} from '@ag.ds-next/react/core';
-import { ModalTitle } from '@ag.ds-next/react/src/modal/ModalTitle';
-import { CloseIcon } from '@ag.ds-next/react/icon';
 import { ModalCover } from '@ag.ds-next/react/src/modal/ModalCover';
 import { Checkbox, ControlGroup } from '@ag.ds-next/react/control-input';
 import { FilterAssigneeSelect } from './FilterAssigneeSelect';
 import { FilterStateSelect } from './FilterStateSelect';
+import { FilterModal, LockScroll } from './FilterModal';
 
-export type ModalDialogProps = PropsWithChildren<{
-	/** The actions to display at the bottom of the modal panel. Typically a `ButtonGroup`. */
-	actions?: ReactNode;
-	/** Function to be called when the modal is closed to accept changes. */
-	onApply: () => void;
-	/** Function to be called when the modal is closed. */
-	onDismiss: () => void;
-	/** The title of the modal dialog. It can span lines but should not be too long. */
-	title: string;
-}>;
-
-const AnimatedStack = animated(Stack);
-
-export const ModalDialog = ({
-	children,
-	title,
-	onApply,
-	onDismiss,
-}: ModalDialogProps) => {
-	// const { titleId } = useModalId()
-	const titleId = 'modal-title';
-
-	const prefersReducedMotion = usePrefersReducedMotion();
-	const animationStyles = useSpring({
-		from: { y: 20, opacity: 0 },
-		to: { y: 0, opacity: 1 },
-		immediate: prefersReducedMotion,
-	});
-
-	return (
-		<FocusLock returnFocus>
-			<AnimatedStack
-				role="dialog"
-				aria-modal="true"
-				background="body"
-				aria-labelledby={titleId}
-				rounded
-				focus
-				maxWidth={tokens.maxWidth.modalDialog}
-				css={{
-					position: 'relative',
-					margin: '0 auto',
-					minHeight: '100vh',
-					[tokens.mediaQuery.min.sm]: {
-						margin: `${mapSpacing(1)} auto ${mapSpacing(1)}`,
-						height: 'calc(100vh - 2rem)',
-						minHeight: 'auto',
-					},
-				}}
-				style={animationStyles}
-			>
-				<Box padding={1.5} borderBottom>
-					<ModalTitle id={titleId}>{title}</ModalTitle>
-				</Box>
-
-				<Box
-					css={{
-						overflowY: 'scroll',
-					}}
-				>
-					{children}
-				</Box>
-
-				<Box borderTop padding={1.5} css={{ marginTop: 'auto' }}>
-					<ButtonGroup>
-						<Button onClick={onApply}>Apply filters</Button>
-						<Button variant="tertiary" onClick={onDismiss}>
-							Cancel
-						</Button>
-					</ButtonGroup>
-				</Box>
-
-				<Button
-					variant="tertiary"
-					aria-label="Close modal"
-					onClick={onDismiss}
-					iconAfter={CloseIcon}
-					css={{
-						position: 'absolute',
-						top: mapSpacing(1),
-						right: mapSpacing(1),
-					}}
-				>
-					Close
-				</Button>
-			</AnimatedStack>
-		</FocusLock>
-	);
-};
-
-export const FilterModal = ({
+export const DashboardFilterModal = ({
 	isOpen,
 	onDismiss,
 	filters,
@@ -196,7 +89,7 @@ export const FilterModal = ({
 		<Fragment>
 			<LockScroll />
 			<ModalCover ref={coverRef}>
-				<ModalDialog onDismiss={onDismiss} onApply={handleApply} title={title}>
+				<FilterModal onDismiss={onDismiss} onApply={handleApply} title={title}>
 					<Stack gap={1} padding={1.5}>
 						<FilterStatusSelect
 							filters={selectedFilters}
@@ -228,11 +121,9 @@ export const FilterModal = ({
 							<Checkbox checked={false}>Mail</Checkbox>
 						</ControlGroup>
 					</Stack>
-				</ModalDialog>
+				</FilterModal>
 			</ModalCover>
 		</Fragment>,
 		document.body
 	);
 };
-
-const LockScroll = () => <Global styles={{ body: { overflow: 'hidden' } }} />;
