+++ .storybook/stories/TableFiltering/components/FilterModal/index.tsx
@@ -1,0 +1,1 @@
+export { FilterModal, LockScroll } from './FilterModal';
